Changed word placements
The blinking "insert coins" at the bottom of the screen did not appear until a coin was inserted. This was remedied by changing it to a variant of "thanks for paying"; since it displays after inserting a quarter, and I haven't found a way to change it before or after. 
With this update, the "please pay up" statement is now unblinking and is placed directly under the title screen, and does not disappear after payment.